% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animate_ggboard.R
\name{animate_ggboard}
\alias{animate_ggboard}
\title{animate_ggboard}
\usage{
animate_ggboard(board_ggplot)
}
\arguments{
\item{board_ggplot}{A object of class \code{ggplot2} as produced by \code{gameRlife::build_ggboard}.}
}
\value{
Used for its side-effect of animating in
}
\description{
Animates the Game of Life board for a given \code{gameRlife::game_of_life} output using \code{gganimate} and \code{gifski}: intended to be viewed in the \code{RStudio} \code{Viewer} tab.
}
\author{
Alex Kalinka, \email{alex.t.kalinka@gmail.com}
}
