% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animate_gol.R
\name{animate_gol}
\alias{animate_gol}
\title{animate_gol}
\usage{
animate_gol(gol, from = NULL, to = NULL)
}
\arguments{
\item{gol}{An object of class \code{gameRlife} as produced by \code{gameRlife::game_of_life}.}

\item{from}{An integer specifying the generation to start at. If \code{NULL} (default) start at generation 0.}

\item{to}{An integer specifying the generation to run to. If \code{NULL} (default) run until the last generation.}
}
\value{
A GIF animation to be viewed in the \code{RStudio} \code{Viewer} tab.
}
\description{
Animates a \code{gameRlife} object.
}
\author{
Alex Kalinka, \email{alex.t.kalinka@gmail.com}
}
