% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/board_df.R
\name{board_df}
\alias{board_df}
\title{board_df}
\usage{
board_df(board, board_size, generation)
}
\arguments{
\item{board}{An \eqn{n^2} binary matrix with living cells indicated by 1s.}

\item{board_size}{An integer specifying the size of the 2D array.}

\item{generation}{A positive integer specifying the number of the focal generation.}
}
\value{
A data frame containing \code{row} and \code{column} columns, as well as the generation number.
}
\description{
Returns the board as a data frame containing co-ordinates of the living cells.
}
\author{
Alex Kalinka, \email{alex.t.kalinka@gmail.com}
}
