% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_gganimate_board.R
\name{build_gganimate_board}
\alias{build_gganimate_board}
\title{build_gganimate_board}
\usage{
build_gganimate_board(board_ggplot, num_gen)
}
\arguments{
\item{board_ggplot}{A object of class \code{ggplot2} as produced by \code{gameRlife::build_ggboard}.}

\item{num_gen}{An integer specifying the number of generations.}
}
\value{
Used for its side-effect of animating in
}
\description{
Returns a \code{gganimate} object of the Game of Life board for a given \code{gameRlife::game_of_life} output using \code{gganimate} and \code{gifski}: intended to be viewed in the \code{RStudio} \code{Viewer} tab.
}
\author{
Alex Kalinka, \email{alex.t.kalinka@gmail.com}
}
