% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_ggboard.R
\name{build_ggboard}
\alias{build_ggboard}
\title{build_ggboard}
\usage{
build_ggboard(board_df)
}
\arguments{
\item{board_df}{A data frame containing the co-ordinates of living cells, as produced by \code{gameRlife::board_df}.}
}
\value{
An object of class \code{ggplot}.
}
\description{
Creates a \code{ggplot2} object of the Game of Life board for a data frame of 1 or more generations.
}
\author{
Alex Kalinka, \email{alex.t.kalinka@gmail.com}
}
