% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coords_living.R
\name{coords_living}
\alias{coords_living}
\title{coords_living}
\usage{
coords_living(board)
}
\arguments{
\item{board}{An \eqn{n^2} binary matrix with living cells indicated by 1s.}
}
\value{
An nx2 matrix with row co-ordinates in the first column and column co-ordinates in the second.
}
\description{
Returns the board co-ordinates of all the living cells.
}
\author{
Alex Kalinka, \email{alex.t.kalinka@gmail.com}
}
