% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/torify_coordinates.R
\name{.torify_coordinates}
\alias{.torify_coordinates}
\title{torify_coordinates}
\usage{
.torify_coordinates(coords, array_max)
}
\arguments{
\item{coords}{An 8x2 matrix of co-ordinates.}

\item{array_max}{An integer specifying the maximum size of the array.}
}
\value{
An 8x2 matrix in which the first column contains row co-oridnates and the second column contains column co-ordinates.
}
\description{
Wraps a grid array into a torus at the edges.
}
\author{
Alex Kalinka, \email{alex.t.kalinka@gmail.com}
}
