% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seed_random.R
\name{make_seed_random}
\alias{make_seed_random}
\title{make_seed_random}
\usage{
make_seed_random(board_size, proportion_living = 0.1)
}
\arguments{
\item{board_size}{An integer specifying the size of the 2D array.}

\item{proportion_living}{A real number in range (0,1). Defaults to 0.1.}
}
\value{
A matrix with \code{proportion_living}*\code{board_size^2} rows and 2 columns providing the co-ordinates of the random seed living cells (rows in the first column, columns in the second).
}
\description{
Creates a random seed for \code{gameRlife::game_of_life} for a given board size and proportion of living cells.
}
\author{
Alex Kalinka, \email{alex.t.kalinka@gmail.com}
}
