% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seed_random.R
\name{seed_random}
\alias{seed_random}
\title{seed_random}
\usage{
seed_random(board_size)
}
\arguments{
\item{board_size}{An integer specifying the size of the 2D array.}

\item{proportion_living}{A real number in range (0,1).}
}
\value{
A matrix with \code{percent_living}*\code{board_size^2} rows and 2 columns providing the co-ordinates of the random seed living cells.
}
\description{
Creates a random seed for \code{gameRlife::game_of_life} for a given board size and proportion of living cells.
}
\author{
Alex Kalinka, \email{alex.t.kalinka@gmail.com}
}
