% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{seed_random4}
\alias{seed_random4}
\title{seed_random4}
\description{
A random seed for the \code{gameRlife::game_of_life}: minimum board size of 50.
}
\details{
A matrix of co-ordinates. Minimum board size of 50.
}
\keyword{data}
