% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_living_neighbours.R
\name{sum_living_neighbours}
\alias{sum_living_neighbours}
\title{sum_living_neighbours}
\usage{
sum_living_neighbours(board, torify = TRUE)
}
\arguments{
\item{board}{An \eqn{n^2} binary matrix with living cells indicated by 1s.}

\item{torify}{Logical indicating whether the board should be wrapped up into a torus to prevent inconsistencies at the edges (defaults to \code{TRUE}).}
}
\value{
An \eqn{n^2} matrix containing the number of living neighbours that each cell has.
}
\description{
Returns a matrix containing the number of living neighbours for each cell on the board.
}
\author{
Alex Kalinka, \email{alex.t.kalinka@gmail.com}
}
