% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_board.R
\name{update_board}
\alias{update_board}
\title{update_board}
\usage{
update_board(board, torify = TRUE)
}
\arguments{
\item{board}{An \eqn{n^2} binary matrix with living cells indicated by 1s.}

\item{torify}{Logical indicating whether the board should be wrapped up into a torus to prevent inconsistencies at the edges (defaults to \code{TRUE}).}
}
\value{
An updated \eqn{n^2} binary matrix with living cells indicated by 1s.
}
\description{
Updates the living and dead cells according to the Game of Life rules.
}
\author{
Alex Kalinka, \email{alex.t.kalinka@gmail.com}
}
